add rotate support to fcs, and others

+++ src/model/tools/myrandom.hpp
@@ -12,7 +12,7 @@ namespace carphymodel {
  */
 inline double rand() {
     static auto seed = std::chrono::system_clock::now().time_since_epoch().count();
-    static std::default_random_engine e(seed);
+    static std::default_random_engine e((unsigned int)seed);
     static std::uniform_real_distribution<double> u(0, 1.);
     return u(e);
 }
